Removed broad exception clause to catch exceptions which may be relevant.
For example, a ResponseError indicating the wrong number of arguments is
raised when the Redis server version is too old to handle the `nx` or `px`
arguments. This exception information is useful, and the case of acquiring
or not acquiring the lock can be handled by expecting either True or None
from the call to `set`.